Remove compiler warnings on 64-bit platform

+++ mxflib/mdtraits.cpp
@@ -995,7 +995,7 @@ void MDTraits_BasicArray::SetString(MDValuePtr Object, std::string Val)
 		if(it == Object->end()) 
 		{
 			// DRAGONS: This will fail for arrays of more that 4 billion entries
-			Object->Resize(Object->size() + 1);
+			Object->Resize(static_cast<UInt32>(Object->size() + 1));
 			it = Object->end();
 			it--;
 		}
